TargetType 위치 변경 및 alamofire upload 혹은 request 경우의 수 분기처리 , 멀티파트 생성 함수 추가
