--- DaangnMarket/DaangnMarket/Scene/Network/TargetType.swift
@@ -1,57 +1,0 @@
-//
-//  TargetType.swift
-//  DaangnMarket
-//
-//  Created by 김지현 on 2022/05/31.
-//
-
-import Foundation
-import Alamofire
-
-protocol TargetType: URLRequestConvertible {
-    var baseURL: String { get }
-    var method: HTTPMethod { get }
-    var path: String { get }
-    var parameters: RequestParams { get }
-}
-
-extension TargetType {
-
-    func asURLRequest() throws -> URLRequest {
-        let url = try baseURL.asURL()
-        var urlRequest = try URLRequest(url: url.appendingPathComponent(path), method: method)
-        urlRequest.setValue(ContentType.json.rawValue, forHTTPHeaderField: HTTPHeaderField.contentType.rawValue)
-
-        switch parameters {
-        case .query(let request):
-            let params = request?.toDictionary() ?? [:]
-            let queryParams = params.map { URLQueryItem(name: $0.key, value: "\($0.value)") }
-            var components = URLComponents(string: url.appendingPathComponent(path).absoluteString)
-            components?.queryItems = queryParams
-            urlRequest.url = components?.url
-        case .body(let request):
-            let params = request?.toDictionary() ?? [:]
-            urlRequest.httpBody = try JSONSerialization.data(withJSONObject: params, options: [])
-        case .requestParameters(let request):
-            let params = request
-            urlRequest.httpBody = try JSONSerialization.data(withJSONObject: params, options: [])
-        }
-
-        return urlRequest
-    }
-}
-
-enum RequestParams {
-    case query(_ parameter: Codable?)
-    case body(_ parameter: Codable?)
-    case requestParameters(_ parameter: [String : Any])
-}
-
-extension Encodable {
-    func toDictionary() -> [String: Any] {
-        guard let data = try? JSONEncoder().encode(self),
-              let jsonData = try? JSONSerialization.jsonObject(with: data),
-              let dictionaryData = jsonData as? [String: Any] else { return [:] }
-        return dictionaryData
-    }
-}
